requirements.txt
nonebot[scheduler]~=1.6.0
aiocqhttp>=1.2
lxml>=4.4.1
pytz>=2019.3
requests>=2.22.0
sogou_tr_free>=0.0.6
zhconv>=1.4.0
TwitterAPI>=2.5.10
matplotlib>=3.2.0
numpy>=1.18.0
beautifulsoup4>=4.9.0
pygtrie>=2.0
tinydb>=4.0
aiohttp>=3.6
peony-twitter[all]~=1.1.7

wmi
openpyxl
xlrd
posix
demjson
websocket-client
Brotli
psutil
feedparser~=5.2.1
pyquery~=1.4.3
httpx~=0.12.1
emoji~=0.5.4
googletrans~=2.4.0
Pillow~=8.0.1
uuid~=1.30
pyyaml
sqlitedict
pixivpy
wordcloud
jieba
opencv-python